@@ -8,7 +8,7 @@ authors = [
     {name = "Kedro"}
 ]
 description = "Kedro-Datasets is where you can find all of Kedro's data connectors."
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = {text = "Apache Software License (Apache 2.0)"}
 dependencies = [
     "kedro>=0.19.7",
@@ -17,13 +17,13 @@ dependencies = [
 dynamic = ["readme", "version"]
 
 [project.optional-dependencies]
-pandas-base = ["pandas>=1.3, <3.0",]
-spark-base = ["pyspark>=2.2, <4.0",]
-hdfs-base = ["hdfs>=2.5.8, <3.0",]
-s3fs-base = ["s3fs>=2021.04",]
-polars-base = ["polars>=0.18.0",]
+pandas-base = ["pandas>=1.3, <3.0"]
+spark-base = ["pyspark>=2.2, <4.0"]
+hdfs-base = ["hdfs>=2.5.8, <3.0"]
+s3fs-base = ["s3fs>=2021.4"]
+polars-base = ["polars>=0.18.0"]
 plotly-base = ["plotly>=4.8.0, <6.0"]
-delta-base = ["delta-spark>=1.0, <4.0",]
+delta-base = ["delta-spark>=1.0, <4.0"]
 networkx-base = ["networkx~=2.4"]
 
 # Individual Datasets
@@ -43,7 +43,7 @@ databricks = ["kedro-datasets[databricks-managedtabledataset]"]
 geopandas-geojsondataset = ["geopandas>=0.6.0, <1.0", "pyproj~=3.0"]
 geopandas = ["kedro-datasets[geopandas-geojsondataset]"]
 
-holoviews-holoviewswriter = ["holoviews~=1.13.0"]
+holoviews-holoviewswriter = ["holoviews>=1.13.0"]
 holoviews = ["kedro-datasets[holoviews-holoviewswriter]"]
 
 huggingface-hfdataset = ["datasets", "huggingface_hub"]
@@ -118,7 +118,7 @@ pandas = [
 pickle-pickledataset = ["compress-pickle[lz4]~=2.1.0"]
 pickle = ["kedro-datasets[pickle-pickledataset]"]
 
-pillow-imagedataset = ["Pillow~=9.0"]
+pillow-imagedataset = ["Pillow>=9.0"]
 pillow = ["kedro-datasets[pillow-imagedataset]"]
 
 plotly-htmldataset = ["kedro-datasets[plotly-base]"]
@@ -154,10 +154,10 @@ spark = [
     spark-sparkstreamingdataset]"""
 ]
 
-svmlight-svmlightdataset = ["scikit-learn>=1.0.2", "scipy~=1.7.3"]
+svmlight-svmlightdataset = ["scikit-learn>=1.0.2", "scipy>=1.7.3"]
 svmlight = ["kedro-datasets[svmlight-svmlightdataset]"]
 
-tensorflow-tensorflowmodeldataset = ["tensorflow~=2.0; platform_system != 'Darwin' or platform_machine != 'arm64'", "tensorflow-macos~=2.0; platform_system == 'Darwin' and platform_machine == 'arm64'",]
+tensorflow-tensorflowmodeldataset = ["tensorflow~=2.0; platform_system != 'Darwin' or platform_machine != 'arm64'", "tensorflow-macos~=2.0; platform_system == 'Darwin' and platform_machine == 'arm64'"]
 tensorflow = ["kedro-datasets[tensorflow-tensorflowmodeldataset]"]
 
 text-textdataset = []
@@ -178,7 +178,7 @@ langchain-chatopenaidataset = ["langchain-openai~=0.1.7"]
 langchain-openaiembeddingsdataset = ["langchain-openai~=0.1.7"]
 langchain-chatanthropicdataset = ["langchain-anthropic~=0.1.13", "langchain-community~=0.2.0"]
 langchain-chatcoheredataset = ["langchain-cohere~=0.1.5", "langchain-community~=0.2.0"]
-langchain = ["kedro-datasets[langchain-chatopenaidataset,langchain-openaiembeddingsdataset,langchain-chatanthropicdataset,langchain-chatcoheredataset ]"]
+langchain = ["kedro-datasets[langchain-chatopenaidataset,langchain-openaiembeddingsdataset,langchain-chatanthropicdataset,langchain-chatcoheredataset]"]
 
 netcdf-netcdfdataset = ["h5netcdf>=1.2.0","netcdf4>=1.6.4","xarray>=2023.1.0"]
 netcdf = ["kedro-datasets[netcdf-netcdfdataset]"]
@@ -227,8 +227,7 @@ test = [
     "jupyterlab>=3.0",
     "jupyter~=1.0",
     "lxml~=4.6",
-    "matplotlib>=3.0.3, <3.4; python_version < '3.10'",  # 3.4.0 breaks holoviews
-    "matplotlib>=3.5, <3.6; python_version >= '3.10'",
+    "matplotlib>=3.5, <3.6",
     "memory_profiler>=0.50.0, <1.0",
     "moto==5.0.0",
     "mypy~=1.0",
@@ -237,7 +236,7 @@ test = [
     "openpyxl>=3.0.3, <4.0",
     "pandas-gbq>=0.12.0",
     "pandas>=2.0",
-    "Pillow~=9.0",
+    "Pillow~=10.0",
     "plotly>=4.8.0, <6.0",
     "polars[xlsx2csv, deltalake]~=0.18.0",
     "pre-commit>=2.9.2",
