@@ -51,7 +51,7 @@ def dummy_graph_data():
 
 class TestGraphMLDataset:
     def test_save_and_load(self, graphml_dataset, dummy_graph_data):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         graphml_dataset.save(dummy_graph_data)
         reloaded = graphml_dataset.load()
         assert dummy_graph_data.nodes(data=True) == reloaded.nodes(data=True)
@@ -60,7 +60,7 @@ class TestGraphMLDataset:
 
     def test_load_missing_file(self, graphml_dataset):
         """Check the error when trying to load missing file."""
-        pattern = r"Failed while loading data from data set GraphMLDataset\(.*\)"
+        pattern = r"Failed while loading data from dataset GraphMLDataset\(.*\)"
         with pytest.raises(DatasetError, match=pattern):
             assert graphml_dataset.load()
 
@@ -100,7 +100,7 @@ class TestGraphMLDataset:
 class TestGraphMLDatasetVersioned:
     def test_save_and_load(self, versioned_graphml_dataset, dummy_graph_data):
         """Test that saved and reloaded data matches the original one for
-        the versioned data set."""
+        the versioned dataset."""
         versioned_graphml_dataset.save(dummy_graph_data)
         reloaded = versioned_graphml_dataset.load()
         assert dummy_graph_data.nodes(data=True) == reloaded.nodes(data=True)
@@ -114,13 +114,13 @@ class TestGraphMLDatasetVersioned:
             versioned_graphml_dataset.load()
 
     def test_exists(self, versioned_graphml_dataset, dummy_graph_data):
-        """Test `exists` method invocation for versioned data set."""
+        """Test `exists` method invocation for versioned dataset."""
         assert not versioned_graphml_dataset.exists()
         versioned_graphml_dataset.save(dummy_graph_data)
         assert versioned_graphml_dataset.exists()
 
     def test_prevent_override(self, versioned_graphml_dataset, dummy_graph_data):
-        """Check the error when attempt to override the same data set
+        """Check the error when attempt to override the same dataset
         version."""
         versioned_graphml_dataset.save(dummy_graph_data)
         pattern = (
