@@ -38,14 +38,14 @@ def dummy_dataframe():
 
 class TestFeatherDataset:
     def test_save_and_load(self, feather_dataset, dummy_dataframe):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         feather_dataset.save(dummy_dataframe)
         reloaded = feather_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded)
 
     def test_exists(self, feather_dataset, dummy_dataframe):
         """Test `exists` method invocation for both existing and
-        nonexistent data set."""
+        nonexistent dataset."""
         assert not feather_dataset.exists()
         feather_dataset.save(dummy_dataframe)
         assert feather_dataset.exists()
@@ -92,7 +92,7 @@ class TestFeatherDataset:
 
     def test_load_missing_file(self, feather_dataset):
         """Check the error when trying to load missing file."""
-        pattern = r"Failed while loading data from data set FeatherDataset\(.*\)"
+        pattern = r"Failed while loading data from dataset FeatherDataset\(.*\)"
         with pytest.raises(DatasetError, match=pattern):
             feather_dataset.load()
 
@@ -154,7 +154,7 @@ class TestFeatherDatasetVersioned:
 
     def test_save_and_load(self, versioned_feather_dataset, dummy_dataframe):
         """Test that saved and reloaded data matches the original one for
-        the versioned data set."""
+        the versioned dataset."""
         versioned_feather_dataset.save(dummy_dataframe)
         reloaded_df = versioned_feather_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded_df)
@@ -166,13 +166,13 @@ class TestFeatherDatasetVersioned:
             versioned_feather_dataset.load()
 
     def test_exists(self, versioned_feather_dataset, dummy_dataframe):
-        """Test `exists` method invocation for versioned data set."""
+        """Test `exists` method invocation for versioned dataset."""
         assert not versioned_feather_dataset.exists()
         versioned_feather_dataset.save(dummy_dataframe)
         assert versioned_feather_dataset.exists()
 
     def test_prevent_overwrite(self, versioned_feather_dataset, dummy_dataframe):
-        """Check the error when attempting to overwrite the data set if the
+        """Check the error when attempting to overwrite the dataset if the
         corresponding feather file for a given save version already exists."""
         versioned_feather_dataset.save(dummy_dataframe)
         pattern = (
