@@ -55,14 +55,14 @@ def json_lines_data(tmp_path):
 
 class TestJSONDataset:
     def test_save_and_load(self, json_dataset, dummy_dataframe):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         json_dataset.save(dummy_dataframe)
         reloaded = json_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded)
 
     def test_exists(self, json_dataset, dummy_dataframe):
         """Test `exists` method invocation for both existing and
-        nonexistent data set."""
+        nonexistent dataset."""
         assert not json_dataset.exists()
         json_dataset.save(dummy_dataframe)
         assert json_dataset.exists()
@@ -117,7 +117,7 @@ class TestJSONDataset:
 
     def test_load_missing_file(self, json_dataset):
         """Check the error when trying to load missing file."""
-        pattern = r"Failed while loading data from data set JSONDataset\(.*\)"
+        pattern = r"Failed while loading data from dataset JSONDataset\(.*\)"
         with pytest.raises(DatasetError, match=pattern):
             json_dataset.load()
 
@@ -208,7 +208,7 @@ class TestJSONDatasetVersioned:
 
     def test_save_and_load(self, versioned_json_dataset, dummy_dataframe):
         """Test that saved and reloaded data matches the original one for
-        the versioned data set."""
+        the versioned dataset."""
         versioned_json_dataset.save(dummy_dataframe)
         reloaded_df = versioned_json_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded_df)
@@ -220,13 +220,13 @@ class TestJSONDatasetVersioned:
             versioned_json_dataset.load()
 
     def test_exists(self, versioned_json_dataset, dummy_dataframe):
-        """Test `exists` method invocation for versioned data set."""
+        """Test `exists` method invocation for versioned dataset."""
         assert not versioned_json_dataset.exists()
         versioned_json_dataset.save(dummy_dataframe)
         assert versioned_json_dataset.exists()
 
     def test_prevent_overwrite(self, versioned_json_dataset, dummy_dataframe):
-        """Check the error when attempting to override the data set if the
+        """Check the error when attempting to override the dataset if the
         corresponding hdf file for a given save version already exists."""
         versioned_json_dataset.save(dummy_dataframe)
         pattern = (
