@@ -35,7 +35,7 @@ def dummy_plot():
 
 class TestJSONDataset:
     def test_save_and_load(self, json_dataset, dummy_plot):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         json_dataset.save(dummy_plot)
         reloaded = json_dataset.load()
         assert dummy_plot == reloaded
@@ -44,14 +44,14 @@ class TestJSONDataset:
 
     def test_exists(self, json_dataset, dummy_plot):
         """Test `exists` method invocation for both existing and
-        nonexistent data set."""
+        nonexistent dataset."""
         assert not json_dataset.exists()
         json_dataset.save(dummy_plot)
         assert json_dataset.exists()
 
     def test_load_missing_file(self, json_dataset):
         """Check the error when trying to load missing file."""
-        pattern = r"Failed while loading data from data set JSONDataset\(.*\)"
+        pattern = r"Failed while loading data from dataset JSONDataset\(.*\)"
         with pytest.raises(DatasetError, match=pattern):
             json_dataset.load()
 
