@@ -155,7 +155,7 @@ class TestEagerParquetDatasetVersioned:
         assert df.shape == (2, 3)
 
     def test_save_and_load(self, versioned_parquet_dataset, dummy_dataframe):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         versioned_parquet_dataset.save(dummy_dataframe)
         reloaded_df = versioned_parquet_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded_df)
@@ -241,7 +241,7 @@ class TestEagerParquetDatasetVersioned:
 
 class TestEagerIPCDatasetVersioned:
     def test_save_and_load(self, versioned_ipc_dataset, dummy_dataframe):
-        """Test saving and reloading the data set."""
+        """Test saving and reloading the dataset."""
         versioned_ipc_dataset.save(dummy_dataframe)
         reloaded_df = versioned_ipc_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded_df)
@@ -345,7 +345,7 @@ class TestEagerCSVDatasetVersioned:
 
     def test_save_and_load(self, versioned_csv_dataset, dummy_dataframe):
         """Test that saved and reloaded data matches the original one for
-        the versioned data set."""
+        the versioned dataset."""
         versioned_csv_dataset.save(dummy_dataframe)
         reloaded_df = versioned_csv_dataset.load()
         assert_frame_equal(dummy_dataframe, reloaded_df)
@@ -446,13 +446,13 @@ class TestEagerCSVDatasetVersioned:
             versioned_csv_dataset.load()
 
     def test_exists(self, versioned_csv_dataset, dummy_dataframe):
-        """Test `exists` method invocation for versioned data set."""
+        """Test `exists` method invocation for versioned dataset."""
         assert not versioned_csv_dataset.exists()
         versioned_csv_dataset.save(dummy_dataframe)
         assert versioned_csv_dataset.exists()
 
     def test_prevent_overwrite(self, versioned_csv_dataset, dummy_dataframe):
-        """Check the error when attempting to override the data set if the
+        """Check the error when attempting to override the dataset if the
         corresponding Generic (csv) file for a given save version already exists."""
         versioned_csv_dataset.save(dummy_dataframe)
         pattern = (
