@@ -94,7 +94,7 @@ class TestDeltaTableDataset:
         catalog = DataCatalog({"delta_in": delta_ds})
         pipeline = modular_pipeline([node(no_output, "delta_in", None)])
         pattern = (
-            r"The following data sets cannot be used with "
+            r"The following datasets cannot be used with "
             r"multiprocessing: \['delta_in'\]"
         )
         with pytest.raises(AttributeError, match=pattern):
